Disable USB power management.
Because it keeps turning off one of my PCIe interface cards which then
the kernel declares dead when it can't power it back on.

Also, it powers off an external hub which has my keyboard plugged into
it, and it also never wakes up.

USB PM seems broken under both 6.8 and 6.11.